1.5.5（本地pdf报告一像素值由2改2.08）

+++ 700sdk/src/main/java/com/Carewell/view/ecg12/EcgPreviewTemplate12Lead12X1.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public class EcgPreviewTemplate12Lead12X1 extends BaseEcgPreviewTemplate {
 
-    public EcgPreviewTemplate12Lead12X1(int width, int height,
+    public EcgPreviewTemplate12Lead12X1(float width, float height,
                                         boolean isDrawGrid, List<String> leadNameList,
                                         float[] gainArray, LeadSpeedType leadSpeedType) {
         this.drawWidth = width;
