+++ 700sdk/src/main/java/com/Carewell/view/ecg12/EcgPreviewTemplate12Lead6X2.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public class EcgPreviewTemplate12Lead6X2 extends BaseEcgPreviewTemplate {
 
-    public EcgPreviewTemplate12Lead6X2(int width, int height,
+    public EcgPreviewTemplate12Lead6X2(float width, float height,
                                        boolean isDrawGrid, List<String> leadNameList,
                                        float[] gainArray, LeadSpeedType leadSpeedType) {
         this.drawWidth = width;
